Base commit for further development and updates
This commit serves as a complete base to all other commits on the stable
and development branch. This commit / release should NOT be used, since
the mod does pretty much nothing right now.

To find the latest updates to the mod, have a look at the development
branch of the repository, or have a look at the GitHub pages for the
project.